refactor: 패키지 위치 리팩토링 ddl -> 각 역할에 맞는 패키지로 전환
