+++ src/main/java/persistence/meta/JavaToJdbcFiledMapper.java
@@ -1,8 +1,9 @@
-package persistence.sql.ddl;
+package persistence.meta;
 
 import java.sql.JDBCType;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import persistence.exception.NotSupertypeException;
 
 public final class JavaToJdbcFiledMapper {
     private static final Map<Class<?>, JDBCType> classTypeMapper = new ConcurrentHashMap<>();
@@ -19,7 +20,7 @@ public final class JavaToJdbcFiledMapper {
     public static JDBCType convert(Class<?> javaType) {
         final JDBCType type = classTypeMapper.get(javaType);
         if (type == null) {
-            throw new IllegalArgumentException("지원 하지 않은 타입 입니다.");
+            throw new NotSupertypeException();
         }
         return type;
     }
