+++ src/test/java/persistence/meta/EntityColumnOptionTest.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.meta;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
